Finished Snakon
Still needs some polish on the animations, but the rest is functioning within unity. Only problem is hitbox not functioning and enemy attacking when the player is behind them but that isn't that difficult to change. Also apparently someone else was working on this at the same time so the github will be screwy.