Tomatroll Unity Implementation
Added in code and animator controller for 3 different attacks. Unsure if the multihitbox is working with them, but the animator choosing one randomly works (as well as corresponding code for cds and anim length). One thing I observed is that he doesn't do the third attack that often, and does the first attack very often, but perhaps that is confirmation bias, because I just used random.range.