Still working on EnemyBehaviorTree
Stunning is confirmed to work now. I think it will work in the middle of an attack too, but we will need the Ham-mer to test that. The final thing that I realized is wrong in the EnemyBehaviorTree script is a logic error concerning the enemy attacking before cd is over. Will fix and then push final version before Monday (hopefully).