Still working
Not done yet, but pushing now because I haven't in a while. All the enemies should be done except benedict and the tomatroll, so they likely will be throwing errors (should still run though). I've done some testing and the new scripts seem to work fine, but especially with sounds we need to make sure those still work.